Implement VerifySolutions for MultipletReduce
modified:   Multiplets.m
If VerifySolutions is to True, then MultipletReduce checks if the
dimensions of input and result are consistent. Specifically, the product
of the dimensions of the factor multiplets must equal the sum of the
dimensions of the decomposition.

For now, we set VerifySolutions -> True by default. In later versions,
when MultipletReduce has been sufficiently tested, the default should
be set to False.